whgrbrg

+++ web/app/plugins/sttv/includes/functions/sttv-webhook-functions.php
@@ -21,7 +21,7 @@ function trial_expiration_checker() {
     if ( !empty( $garbage ) ) {
         foreach ( $garbage as $g ) {
             $umeta = get_user_meta( $g['wp_id'], 'sttv_user_data', true );
-            if ( $g['exp_date'] > 0 && $umeta ) {
+            if ( ( $time > $g['exp_date'] && $g['exp_date'] > 0 ) && $umeta ) {
                 try {
                     $customer = \Stripe\Customer::retrieve( $umeta['customer'] );
                     $customer->delete();
